+++ app/helpers/admin/base_helper.rb
@@ -17,7 +17,7 @@ module Admin::BaseHelper
   def admin_sign_out_path
     case Typus.authentication
     when :devise
-      send("destroy_#{Typus.user_class_name.underscore}_session_path")
+      send("destroy_#{Typus.class.name.underscore}_session_path")
     else
       destroy_admin_session_path
     end
@@ -25,7 +25,7 @@ module Admin::BaseHelper
 
   def admin_edit_user_path(user)
     {
-      controller: "/admin/#{Typus.user_class.to_resource}",
+      controller: "/admin/#{Typus.class.to_resource}",
       action: 'edit',
       id: user.id,
     }